tests/test_hard_and.py
import jax.numpy as jnp

from neurallogic import hard_and, harden


def test_include():
    test_data = [
        [[1.0, 1.0], 1.0],
        [[1.0, 0.0], 0.0],
        [[0.0, 0.0], 1.0],
        [[0.0, 1.0], 1.0],
        [[1.1, 1.0], 1.0],
        [[1.1, 0.0], 0.0],
        [[-0.1, 0.0], 1.0],
        [[-0.1, 1.0], 1.0]
    ]
    for input, expected in test_data:
        assert hard_and.soft_and_include(*input) == expected
        assert hard_and.hard_and_include(*harden.harden(input)) == harden.harden(expected)

def test_neuron():
    test_data = [
        [[1.0, 1.0], [1.0, 1.0], 1.0],
        [[0.0, 0.0], [0.0, 0.0], 1.0],
        [[1.0, 0.0], [0.0, 1.0], 0.0],
        [[0.0, 1.0], [1.0, 0.0], 0.0],
        [[0.0, 1.0], [0.0, 0.0], 1.0],
        [[0.0, 1.0], [1.0, 1.0], 0.0]
    ]
    for input, weights, expected in test_data:
        input = jnp.array(input)
        weights = jnp.array(weights)
        assert jnp.array_equal(hard_and.soft_and_neuron(weights, input), expected)
        assert jnp.array_equal(hard_and.hard_and_neuron(harden.harden(weights), harden.harden(input)), harden.harden(expected))

def test_layer():
    test_data = [
        [[1.0, 0.0], [[1.0, 1.0], [0.0, 1.0], [1.0, 0.0], [0.0, 0.2]], [0.0, 0.0, 1.0, 0.8]],
        [[1.0, 0.4], [[1.0, 1.0], [0.0, 1.0], [1.0, 0.0], [0.0, 0.0]], [0.4, 0.4, 1.0, 1.0]],
        [[0.0, 1.0], [[1.0, 1.0], [0.0, 0.8], [1.0, 0.0], [0.0, 0.0]], [0.0, 1.0, 0.0, 1.0]],
        [[0.0, 0.0], [[1.0, 0.01], [0.0, 1.0], [1.0, 0.0], [0.0, 0.0]], [0.0, 0.0, 0.0, 1.0]]
    ]
    for input, weights, expected in test_data:
        input = jnp.array(input)
        weights = jnp.array(weights)
        expected = jnp.array(expected)
        assert jnp.array_equal(hard_and.soft_and_layer(weights, input), expected)
        assert jnp.array_equal(hard_and.hard_and_layer(harden.harden(weights), harden.harden(input)), harden.harden(expected))

